requirements.txt
numpy
Pillow
tensorflow-gpu
keras
git+https://www.github.com/keras-team/keras-contrib.git
h5py
